# Add symbolic::Variables::operator+=