--- a/common/symbolic_variables.h
+++ b/common/symbolic_variables.h
@@ -137,6 +137,16 @@ class Variables {
 
   friend Variables intersect(const Variables& vars1, const Variables& vars2);
 
+  /**
+   * Updates this Variables with the result of set-union (@p this, @p var).
+   */
+  Variables& operator+=(const Variable& var);
+
+  /**
+   * Updates this Variables with the result of set-union (@p this, @p vars).
+   */
+  Variables& operator+=(const Variables& vars);
+
  private:
   /* Constructs from std::set<Variable>. */
   explicit Variables(std::set<Variable> vars);
@@ -144,12 +154,6 @@ class Variables {
   std::set<Variable> vars_;
 };
 
-/** Updates @p var1 with the result of set-union(@p var1, @p var2). */
-// NOLINTNEXTLINE(runtime/references) per C++ standard signature.
-Variables operator+=(Variables& vars1, const Variables& vars2);
-/** Updates @p vars with the result of set-union(@p vars, { @p var }). */
-// NOLINTNEXTLINE(runtime/references) per C++ standard signature.
-Variables operator+=(Variables& vars, const Variable& var);
 /** Returns set-union of @p var1 and @p var2. */
 Variables operator+(Variables vars1, const Variables& vars2);
 /** Returns set-union of @p vars and {@p var}. */