This replaces operator+=(symbolic::Variables vars1, symbolic::Variables vars) which unnecessarily calls copying inside the function. Speeds up operator+=.

In benchmark_mathematical_program.cc, sos1 time drops from 2.6s to 2.5s, and sos2 drops from 8s to 6.5s on my P15.

As discussed in https://drakedevelopers.slack.com/archives/C3L92BM2Q/p1650590014510329?thread_ts=1649977461.631799&cid=C3L92BM2Q